refactoring suggestion for the example code to make it a python function
The original code was mainly designed to be used for CLI/argparse via command-line. I foresee a need to use it as a standalone python function. So, this refactoring introduced the create_unified_session_metadata() while keeping the main() function at the bottom so it can be used for both usecases.